pali: Implement (lazy) pause for workers in ThreadPool
This changes enables the feature to pause workers in the
ThreadPool.

Currently once we pause a worker , it will stop taking any request(s)
after at most one taks from queue. That is why it is called lazy pause.
Since the worker thread might be already doing the (blocking) wait on
queue for next item, we cannot guarantee immediately stop from taking
further requests unless we kill the thread. If we kill the thread we
lose the ability to re-enable the worker and will only start a new
worker.